refactor(gateway): extract GatewayTurnExecMixin from run.py (slice 29 of #54962)
